+++ geometry/colmesh_field.c
@@ -272,7 +272,7 @@ void colmesh_field_start_exchange(colmesh_field_t* field)
   START_FUNCTION_TIMER();
 
   // Start the xy exchange.
-  int stride = field->num_components;
+  int stride = (int)field->num_components;
   field->ex_token = exchanger_start_exchange(field->ex, field->buffer, stride, 0, MPI_REAL_T);
   STOP_FUNCTION_TIMER();
 }
